internal/llsync: add low-level synchronization package
This adds an internal/llsync package that provides low level
synchronisation primitives for the runtime and sync packages.
The atomic types are currently not atomic, but since they share the same
API with sync/atomic we can trivially add some type aliases under a
different build tag in the future when real atomicity is needed for
threading.